# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ShopWise API
  version: 2.0.0
  description: |
    Contrat de l'API backend ShopWise pour la gestion des commerces, clients,
    prestations, rendez-vous, comptes clients et programmes de fidelite.

    Les endpoints metier commercants sont proteges par authentification.
    Les endpoints commercants utilisent une session serveur. Le navigateur doit
    envoyer les cookies (`withCredentials: true`). Les requetes d'ecriture doivent
    renvoyer le cookie `XSRF-TOKEN` dans l'en-tete `X-XSRF-TOKEN`.

    En local, le bootstrap cree `owner@shopwise.local` avec le mot de passe
    `Shopwise123!` lorsque la base ne contient encore aucun compte commercant.
servers:
  - url: http://localhost:8080
    description: Environnement local
...
security:
  - merchantSession: []
paths:
...
  /api/merchant-invitations/{token}/accept:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Business team]
      operationId: acceptMerchantInvitation
      summary: Accepter une invitation
      description: |
        Pour une nouvelle personne, firstName, lastName et password sont obligatoires.
        Si l'email appartient deja a un compte, ce compte doit etre connecte.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantInvitationAcceptRequest"
      responses:
        "200":
          description: Invitation acceptee et session ouverte
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantSessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /api/merchant-invitations/{id}/resend:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      tags: [Business team]
      operationId: resendMerchantInvitation
      summary: Renvoyer une invitation avec un nouveau jeton
      responses:
        "200":
          description: Invitation renvoyee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantInvitationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/merchant-invitations/{id}/revoke:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      tags: [Business team]
      operationId: revokeMerchantInvitation
      summary: Revoquer une invitation en attente
      responses:
        "200":
          description: Invitation revoquee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantInvitationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/merchants:
    get:
      tags: [Merchants]
      operationId: listMerchants
      summary: Lister les commercants
      deprecated: true
      parameters:
        - $ref: "#/components/parameters/BusinessIdQuery"
      responses:
        "200":
          description: Liste des commercants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MerchantResponse"
    post:
      tags: [Merchants]
      operationId: createMerchant
      summary: Creer un commercant
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantRequest"
      responses:
        "201":
          description: Commercant cree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/merchants/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      tags: [Merchants]
      operationId: getMerchant
      summary: Consulter un commercant
      deprecated: true
      responses:
        "200":
          description: Commercant trouve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Merchants]
      operationId: updateMerchant
      summary: Mettre a jour un commercant
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantRequest"
      responses:
        "200":
          description: Commercant mis a jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/clients/{clientId}/loyalty:
    parameters:
      - $ref: "#/components/parameters/ClientId"
    get:
      tags: [Loyalty]
      operationId: getLoyaltyAccount
      summary: Consulter le solde de fidelite d'un client
      responses:
        "200":
          description: Compte fidelite trouve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltyAccountResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Loyalty]
      operationId: createLoyaltyAccount
      summary: Creer le compte fidelite d'un client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoyaltyAccountRequest"
      responses:
        "201":
          description: Compte fidelite cree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltyAccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/clients/{clientId}/loyalty/credit:
    parameters:
      - $ref: "#/components/parameters/ClientId"
    post:
      tags: [Loyalty]
      operationId: creditLoyaltyPoints
      summary: Crediter manuellement des points
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoyaltyPointsRequest"
      responses:
        "200":
          description: Nouveau solde
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltyAccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/clients/{clientId}/loyalty/debit:
    parameters:
      - $ref: "#/components/parameters/ClientId"
    post:
      tags: [Loyalty]
      operationId: debitLoyaltyPoints
      summary: Debiter manuellement des points
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoyaltyPointsRequest"
      responses:
        "200":
          description: Nouveau solde
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltyAccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/clients/{clientId}/loyalty/transactions:
    parameters:
      - $ref: "#/components/parameters/ClientId"
    get:
      tags: [Loyalty]
      operationId: listLoyaltyTransactions
      summary: Consulter l'historique de fidelite
      responses:
        "200":
          description: Transactions de la plus recente a la plus ancienne
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoyaltyTransactionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    merchantSession:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Session serveur du compte commercant.
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
    ClientId:
      name: clientId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
    BusinessIdQuery:
      name: businessId
      in: query
      required: true
      description: Commerce actif selectionne par le compte connecte.
      schema:
        $ref: "#/components/schemas/Id"
  responses:
    BadRequest:
      description: Requete invalide ou operation metier interdite
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Session absente, identifiants invalides ou compte inactif
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Le compte connecte n'a pas acces au commerce ou a l'operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Une ressource equivalente existe deja
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Gone:
      description: Le jeton d'invitation a expire
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Id:
      type: integer
      format: int64
      minimum: 1
      example: 1
    LocalDateTime:
      type: string
      description: Date et heure locale ISO-8601 sans fuseau horaire.
      example: "2026-07-20T09:30:00"
...
    TransactionType:
      type: string
      enum: [EARNED, REDEEMED, ADJUSTMENT]
    MembershipRole:
      type: string
      enum: [OWNER, MANAGER, STAFF]
    InvitationStatus:
      type: string
      enum: [PENDING, ACCEPTED, REVOKED, EXPIRED]
...
    MerchantAccountResponse:
      type: object
      required: [id, firstName, lastName, email, active]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        active:
          type: boolean
        lastLogin:
          allOf:
            - $ref: "#/components/schemas/LocalDateTime"
          nullable: true
    BusinessMembershipResponse:
      type: object
      required: [id, businessId, businessName, accountId, firstName, lastName, email, role, active]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        businessId:
          $ref: "#/components/schemas/Id"
        businessName:
          type: string
        accountId:
          $ref: "#/components/schemas/Id"
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/MembershipRole"
        active:
          type: boolean
    MerchantSessionResponse:
      type: object
      required: [account, businesses]
      properties:
        account:
          $ref: "#/components/schemas/MerchantAccountResponse"
        businesses:
          type: array
          items:
            $ref: "#/components/schemas/BusinessMembershipResponse"
...
    MerchantInvitationAcceptRequest:
      type: object
      description: Champs requis uniquement lorsqu'aucun compte n'existe encore pour l'email invite.
      properties:
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
          format: password
          minLength: 8
          writeOnly: true
    MerchantInvitationResponse:
      type: object
      required: [id, businessId, businessName, email, role, status, token, expiresAt, createdById, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        businessId:
          $ref: "#/components/schemas/Id"
        businessName:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/MembershipRole"
        status:
          $ref: "#/components/schemas/InvitationStatus"
        token:
          type: string
          format: uuid
          description: Jeton a transmettre manuellement au destinataire.
        expiresAt:
          $ref: "#/components/schemas/LocalDateTime"
        createdById:
          $ref: "#/components/schemas/Id"
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
...
    MerchantRequest:
      type: object
      required: [businessId, firstName, lastName, email, role]
      properties:
        businessId:
          $ref: "#/components/schemas/Id"
        firstName:
          type: string
          minLength: 1
          example: Marie
        lastName:
          type: string
          minLength: 1
          example: Dupont
        email:
          type: string
          format: email
          example: marie@chez-marie.fr
        role:
          type: string
          minLength: 1
          example: OWNER
        active:
          type: boolean
          default: true
    MerchantResponse:
      allOf:
        - type: object
          required: [id]
          properties:
            id:
              $ref: "#/components/schemas/Id"
        - $ref: "#/components/schemas/MerchantRequest"
...
    LoyaltyAccountRequest:
      type: object
      properties:
        pointsBalance:
          type: integer
          format: int32
          minimum: 0
          default: 0
    LoyaltyPointsRequest:
      type: object
      required: [points]
      properties:
        points:
          type: integer
          format: int32
          minimum: 1
          example: 10
    LoyaltyAccountResponse:
      type: object
      required: [id, clientId, pointsBalance, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        clientId:
          $ref: "#/components/schemas/Id"
        pointsBalance:
          type: integer
          format: int32
          minimum: 0
          example: 25
        updatedAt:
          $ref: "#/components/schemas/LocalDateTime"
    LoyaltyTransactionResponse:
      type: object
      required: [id, type, pointsDelta, reason, transactionDate]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        appointmentId:
          allOf:
            - $ref: "#/components/schemas/Id"
          nullable: true
        type:
          $ref: "#/components/schemas/TransactionType"
        pointsDelta:
          type: integer
          format: int32
          description: Valeur positive pour un credit et negative pour un debit.
          example: 25
        reason:
          type: string
          example: Appointment honored
        transactionDate:
          $ref: "#/components/schemas/LocalDateTime"
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Resource not found