# src/main/java/com/example/shopwiseapi/business/MerchantRequest.java
package com.example.shopwiseapi.business;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;

public record MerchantRequest(
        @NotNull Long businessId,
        @NotBlank String firstName,
        @NotBlank String lastName,
        @NotBlank @Email String email,
        @NotBlank String role,
        Boolean active
) {
}